Major fix for Python 2.6: fix db cursor speed like 2.5
In python 2.6, the hasattr() will undergo subclass checks. So, it is
unwise to call it for the "object". That subclass check would overflow
the recursion limit, on each cursor delete operation, which ends up in
wasted CPU (and slow OpenObject responsiveness).

Now, hopefully, OpenObject will run in 2.6 as fast as it ran in 2.5 .

bzr revid: p_christ@hol.gr-20090420172016-1dgiys0cj90efh34