[IMP] Partial restrict operator for domains.
In some cases, when the domain is restricted based on an optional field,
we want it to match anything if that field is not set. There, use the
'=?' operator to match "left = right OR right IS NULL/False".

eg. match the production lot among all products, when product is not
specified (helps barcode scanners pick up the serial no.).

bzr revid: p_christ@hol.gr-20090709181718-dk3ykimzddl9l07o