document: fix content indexer's handling of the "file" command
Content indexer spawns the *nix "file" command in an attempt to determine
the type of arbitrary files. The way we have fetched the stdout of the
child process had been flawed, leaving a zombie process for every file
scanned. It was not a matter of closing the file descriptors, rather an
issue of cleaning up the child process.

bzr revid: p_christ@hol.gr-20101209112607-psu98d0h2e7xqwu7