account: fix translatability of the general ledger, in landscape
That report is a separate instance of the general ledger report, with an
orm name of "account.general.ledger_landscape". If the ir.action.report.xml
is not given a reference name in ir.model.data, the landscape report
wouldn't be exported under that res_id for translations.

bzr revid: p_christ@hol.gr-20110105120814-d5o8be3t3r2500xn