[IMP] ir.rule: local rules will be combined with OR within the same group
This makes it more intuitive: global rules are restrictions and
always combined with AND, while local rules are always combined
with OR, regardless if they are in the same group or different
groups. It makes it also easier to add rules in an existing group
to grant more rights.

bzr revid: odo@openerp.com-20110601130156-c0swo9w21zqun2gl