[IMP] fields.function: don't wrap int values unless really needed
There was no need to wrap/cast int values for values less
than 2^31-1, which the highest XML-RPC int value.
Also, instead of wrapping them in strings, we can use
float values, which are 64bits based in the XMLRPC standard,
and closer to the real value (for comparisons, etc.).
Added note for integer_big, as a reminder of this possible
issue.

bzr revid: qdp-launchpad@openerp.com-20110818120550-ulvffm6ka9f3c5ym