[FIX] fix useless and invalid casts in domain fields
The ORM will automatically append 00:00:00 to a date
literal that is being compared with a timestamp field
with operators '>' or '>=', and 23:59:59 if compared
with operators '<' or '<='.
So a trivial domain for 'Today' is:
[('tstamp_field','>=',current_date),('tstamp_field','<=',current_date)]
taking advantage of the current_date helper that
is defined as the current date in YYYY-MM-DD format.

bzr revid: odo@openerp.com-20110924005547-pvgblte6r14so7mw