[IMP] email_template: added opt_out flag on res.partner
In the context of email templates or for related features
like marketing campaigns or automated business document
notification, this flag may be used to disable the sending
of template-based emails to the partner.
It will not prevent sending an email based on a template
by default, each feature using a template should implement
this if needed.

bzr revid: odo@openerp.com-20111005113846-qik6d5gxd4xqeqdj