[FIX] m2o values display in list views (mostly under onchange condition)
Normal m2o values are pairs of (id, name) (name_get format, basically)
but in some cases (lazyness & al), APIs (often onchange results) only
return the `id` part, which is not sufficient to display the m2o
as-is, a name_get has to be performed.

Added an intermediate step in displaying a table cell (in list view):
if we're displaying an m2o cell (~a cell for an m2o column) and the
value is a number, then dispatch a name_get and keep trucking (display
the rows without the m2o values).

When the name_get returns, it will simply refresh the row and diplay
the m2o name.

bzr revid: xmo@openerp.com-20111014084310-pv2mmmy28fwwk9sm