[FIX] report_webkit: translations lookup should ignore report name for webkit reports
Webkit/Mako translations are quite different from RML
report translations, and very similar to translations
for Python code terms, i.e. with _(). The method to
lookup the translations at rendering time should thus
be similar to what _() does for Python code terms, and
not depend on the name of the report, which is not
known when exporting the translation template, and is
therefore not available in POT meta-data.
This bears no performance penalty, as we are still
only looking inside 'report'-typed translations with
the matching 'source term'.
The only downside is that this prevents per-report
translations, but this is not supported by our
translation import/export system anyway, so it 
does not matter (msgid strings must be unique
within a PO file, and we don't support context-
sensitive PO files yet).

lp bug: https://launchpad.net/bugs/819334 fixed

bzr revid: odo@openerp.com-20120119102145-tp3w80bhjwq11q84