[REM] project_scrum: remove PO files after disabling template, module is deprecated in 7.0
7.0 alternative is to use the regular project module.
The product backlog will be the first stage in the
project kanban view. The sprint is the kanban view,
and can be customized with specific stages, etc.
The burdown chart is now available as a specific
analysis view in the regular project module.

bzr revid: odo@openerp.com-20120628110619-175y4zvyr0j2cd2p