[FIX] handling of @onwrite
extraction was broken: using _.bind/3 will partially apply the
function (first argument) on top of binding it to a context (second
argument), but the partial application will be from the *left* (which
is pretty much the only one making sense, short of kwargs).

handle_onwrite_record took its arguments the wrong way around: the
partially applied one was specified on the right and the "actually
applied" one was on the left, so it used the wrong argument and ended
up blowing up the rest of the code.

bzr revid: xmo@openerp.com-20120718101744-bbbshq60x2kqhob6