[FIX] mail.thread: fix message_process to avoid duplicate deliveries
Due to a programming error in the code, any message that had
several values in the References header create as many
duplicate copies of the message as they were previous
messages in the discussion.
Also cleaned up and simplified the code.

bzr revid: odo@openerp.com-20120810114838-k273zok10911l6ij