[FIX] base_crypt: avoid table lock on res_users during simple auth check
Performing a full table lock is potentially very slow
due to the number of tables holding foreign keys towards
res_users, so we should avoid doing that for simple
auth check() calls. The point was only to avoid failing
the update of the last login timestamp, which we don't
actually care much about. So we can replace that with a
simple SELECT FOR UPDATE NOWAIT on that specific user
row and and skip the update in case the LOCK was
not granted

bzr revid: odo@openerp.com-20120813143552-3g3z0q9n1iox8a0b