[IMP] tools.html2plaintext: consistent use of lxml.etree.HTMLParser to convert HTML to plaintext
We used to switch to using BeautifulSoup when available, but that
lead to inconsistent behavior depending on the installed Python
packages, and sometimes lead to bad surprises. There is no advantage
in using BeautifulSoup rather than HTMLParser, and the latter is
always available.

bzr revid: odo@openerp.com-20121015120934-njaylf99dc5zekfw