[FIX] db creation: change master password field back to type=password
This is not only dangerous but also
completely inconsistent with the other
db management screens. It also represents
little improvement in terms of usability
for first-time users, since the password
is now mentioned on the comment just above
the field.

bzr revid: odo@openerp.com-20130318155618-z9s2izahnv1j2no6