[FIX] subscription: prevent lockup during cron job execution
Due to the way cron jobs are executed in multi-process
environments, the cron arbiter holds an exclusive lock
on the cron task row in the database.
During the course of the subscription cron jobs execution
they attempt to alter a record that has a FK towards
the cron task row, effectively deadlocking the whole
cron processing thread.
The fix is rather hackish: remove the explicit FK
constraint from the subscription model towards
ir.cron - this prevent PostgreSQL from requesting
an exclusive lock and triggering a deadlock.
This is only a temporary solution, the subscription
module should be rewritten for the next stable release.

bzr revid: odo@openerp.com-20130320153430-bho6ou8b7ziohcb0