[REVERT] account.invoice: remove "commercial entity" domain on partner_id, to keep it consistent everywhere
Being able to choose a contact to invoice is very important
because the invoice must go to the right person.
When the invoice creation is automated by e.g. a Sales Order
any partner can be used (even though the system will
preselect the `invoicing` partner of the company if
one exists).
We have to keep that freedom when manually creating invoices.

bzr revid: odo@openerp.com-20130410135308-pd54c7ml8t3pcefl