[FIX] mail: mail composer: Use a text field instead of a html field
html fields are sanitized just before saving data into the database.
In the composer, this cause the sanitization of the templated
message (for mass mailing messages) before the rendering, forbidding
the use of templated html links.
The html of the message is sanitized by  message_post()

bzr revid: chs@openerp.com-20130411182226-lwnpkh4tmyswjjnc