[FIX] account: restrict bank account post creation for account and journal
Do not create account and journal when there no liquidity account with a parent.
This removes a traceback but may lead to no auto configuration. This is not considered as critical as we prefer that instead of configuring wrongly accounts and journals

bzr revid: mat@openerp.com-20130610110401-1pgpql7r688et8sp