[FIX] disable relative positioning of cke widget wrapper
It's only used to be able to position the drag & drop widget
absolutely, and we've already disabled that one.

Product page is a gory mess of absolute positioning, if product images
are changed into a t-field the relative positioning on the wrapper
makes the image itself disappear completely.

bzr revid: xmo@openerp.com-20131002144554-1rg4zmwu0btf0r0f