[IMP] website_blog: updated models
- removed template field from category. This field was coming from the old document_page module. However
it has been decided to not keep the template feature on categories. Previously there was an onchange
in the document_page form view, allowign to define a template on the page category that will be a default
value for the page content. It has been decided to drop that feature because of the complexity to manage
snippets / html in the category without adding a category editor in the website front-end.
- updated data/demo/views accordingly
- to replace the template feature, it will be possible to duplicate a blog post directly in the front-end.
This will come in a future patch. However the copy method of blog_post has already been fixed to avoid
duplicating website_message_ids, and to reset website_published (_datetime) fields.

bzr revid: tde@openerp.com-20131011125832-dou1ep9ey2ffwa7s