[FIX] m2o t-field rendering: use uid 1 for name_get
This is done automatically on a read(), but t-field uses a browse,
which does not do access-rights bypassing. We've already browsed the
root record so we should have the rights to the reading, and we just
want the textual representation for the child m2o.

bzr revid: xmo@openerp.com-20131025124720-i10eywz70o9c9dpc