[FIX] web: correct date and datetime parsing at end of month.
As `new Date()` return the current date, when we are the 31,
setting the month to a month with less than 31 days will change
the date to next day (1st of next month). This result to a date(time)
object one month ahead of the wanted date(time).

bzr revid: chs@openerp.com-20131031103333-vt68132ptj9sbr04