[FIX] stock: programming error in average price computation for multiple lines of the same product
While processing a picking we must keep track of
previously processed lines as they modify the
stock on hand but are not yet included in the
`qty_available` function.
Negative stock on hand is handled as if
the stock was zero as far as the average
price computation is concerned.

bzr revid: odo@openerp.com-20131104171245-z1lgsplyu4cdz9gc