[REVERT] 9401 dle@openerp.com-20130830125155-1vedifnupu2xvth7
This had as side-effect to not allow splitting BOMs
(a manufacturing order for 1 unit of a BOM producing 10 units consumed the lines like it was 10 units, not 1)

This fix was to avoid having a fraction of a unit (for instance, 0.5 unit).

But, finally, it is preferable to allow splitting units:
1. Most users do not use several uoms, and in this case, the unit is the default uom (hidden).
2. At the moment, it is allowed to ask a manufacturing order splitting up the unit uom
(It is allowed to ask the production of 0.5 USB Adapter, for instance)
Thus, we should allow the splitting up of the unit uom in the lines too.

bzr revid: dle@openerp.com-20140124120102-we2yxio553ws2yz4