[FIX] account_analytic_analysis: currency were wrong when preparing recurring invoices
The id of the sale pricelist of the customer was set as currency_id, instead of using the id of the currency of the pricelist. (Obviously)
Furthermore, the currency of the contract pricelist takes the lead.
Then, if none of these two are set, we choose the currency of the company, finally.

bzr revid: dle@openerp.com-20140204115226-rowmbr5223fcu08k