[FIX] website_event_track: employees should not be restricted to public events only
Also switched rule to noupdate. A manual fix of existing installations
will be necessary, as is usually the case for ir.rules changes.
This module has not made it into a LTS version yet, so it is acceptable.
+ Added a _description for the event.track model to avoid it being
replaced by inherited descriptions.

bzr revid: odo@openerp.com-20140225115742-ghnt3khn6hvycnnw