[FIX] crm: make days to close and days to open simple calendar delta
When computing the duration to close or to open a lead, do not use some kind of complex computation with working days. It does not make sense to use the number of intervals with the number of days. The computation was incorrect and very slow.
Replaced by simple calendar delta.