[FIX] orm: custom m2m with different label
At rev 84e9a67cdf78db94cb7a09543c1b7ac4ad19d8b4 a check to avoid the creation of ir.model.relation for custom modules was added. The condition is not correct as based on the string instead of the field name. We do not have access to column name at this level but the the m2m relation table do start with x_ for custom fields (see __init__ method).