[FIX] web: webclient_bootstrap: render the menus when the user is really a user
The webclient rendering is done in two parts: the server renders the boilerplates
including the menus, and then the  javascript launch the webclient using the
server-side boilerplates. In some case, we want to render the boilerplates but not
the menus (auth_calendar). Added groups around the t-call to the menu in the web
template. Fixes issue #2044.