[FIX] account.invoice.line: move context to last position in onchange methods
The methods product_id_change() and uos_id_change() have been converted to the
new api, and now use the decorator @multi.  When invoked with the old api, by
convention the methods will take the last argument as the context.  But this
will not work properly for those methods, as the context is passed in another
position.  In order to avoid an argument swap in the api wrapper, we moved the
context to its expected position.

Fixes #1943