[IMP] Change documentation + new action_consume/qty/origin corrections
Simplify the action_consume of the consumption lines after the corrections
by Kevin Wang.  Also the UoMs are revised as the action_consume uses the default UoM
of the product.

We have to avoid circular boms where a child bom should not contain the product that
represents the parent bom, but it is possible for example to use another product of the parent bom in
the child bom.

As the consume line move has no procurement rule, its origin will have no description.  So, when there is
none it will also check the description of the previous move (when passed to procurement for example) This way
the chained moves or purchase order for example will have the MO-number as origin and not nothing.

[IMP] Change assignation

[IMP] UoM changes continuation

[IMP] Make sure we can use 2 times the same product in a BoM

[IMP] Source document for consume lines to procurement