[FIX] load view from file shall not raise an exception
In --dev mode, when a view can't be loaded from the file in spite of
being referenced in field `arch_fs`, the current behavior is to raise
an exception. In most cases, developpers don't want the view to fail
so the behavior has been changed in favor of a logger warning and
a fallback to `arch_db`.