[FIX] account: check for profit/loss accounts
in method `button_confirm_cash` of `account.cash.statement`

The check was verifying that the profit/loss account
was set on the journal, and if it was, it raised that it
wasn't, which is obviously wrong.

This was solved in Odoo 8.0 by replacing the code
by something more readable in 9dc9169, and the same
logic to check that the profit/loss accounts are
set is still there.

Closes #2924