[IMP] product: search on internal reference
When searching on the name of a product.template, search also on the default_code. Move the filter from product.product to product.template.
The field default_code is on the product.product but accessible through a related field, meaning when reading it will fetch only the first variant but searching on all variants.
Fixes #3399