[FIX] sale_stock: product_id_change_with_wh on sale order popup too
product_id_change_with_wh onchange was not anymore called when using the popup form in sale order line.

The behavior was different concerning the product onchange when being in the editable list or in the popup form
For instance, for make to stock products not having enough stock, the warning was displayed when being in editable list, but not in the popup form

This is a regression of rev. 86f785ae1b2741ab67384f3022b971ab638daeb8

opw-619624