[FIX] tour.js: remove synchro between selected property and attribute
This line was previously used to allow attribute selectors to match the
`selected` property, by changing the attribute each time the property
was changed. Now it doesn't make sense since rev c611846 introduced
property selectors + it seems to mess with phantomjs, as the test does
not pass in phantomjs environment but passes ok in a "real" webkit browser.