[IMP] add odoo.py shell subcommand
It starts an interactive Python console. If a database is provided, with the
`-d` argument `env` is bound to an Environement with the uid and context of
SUPERUSER_ID and `self` is bound to a RecordSet of the corresponding res.users.

Courtesy of Daniel Reis, reworked to be a bit more agnostic.

Command class is moved from __init__.py to command.py to improve imports of
__init__.py