[FIX] stock_account: avoid getting 'stock.picking()' as reference
In the new api an empty recordset converted to string is the name its class
while previously it was converted to an empty string.

The valid v7 condition sould have probably been
    move.picking_id and move.picking_id.name or False
but in v8, simply move.picking_id.name is enough and avoids getting these
accounting entries strangly named when there is no picking.