[FIX] Make sure we can select dropship on purchase order and see the related customer and invoicing based on shipments works for sale only
[FIX] Sale should lead to invoicing
[IMP] Journal should not depend on picking type code when in dropship

Make sure drop shipment with invoicing on sale orders only, works.
Also the picking type of dropshipping should be incoming.  This
makes it possible to select it on the purchase order.  This also fixes #3421.
And we can create a dropship from a purchase only. (not the good way)
For this, we make the invoicing based on the shipment look at the
purchase related and not on its picking type code as the picking type
of dropshipping is changed back to incoming.
The purchase order will also show the customer address when in
a dropship case.

Closes #3421