[FIX] web: fix breadcrumbs duplicate form views
and prevent the last breadcrumb to be clickable (revert of 61dec05).
The last breadcrumb clickable to refresh the view is no longer necessary
with the form view icon available.

The weird code on the click handler for switching views is due to the
fact that most views refresh with a do_search but the form view is
refreshed with a do_show.  A more generic way to update the view
seems necessary to prevent such artifacts.