[FIX] web: properly hide list view in pop up dialogs
a many2many field can open a pop up dialog with different views, such
as form view and list view.  when the user click on the create button,
it should hide the list view and display the form view.  apparently,
this behavior was hardcoded in the pop up dialog, and broken at some
point between v8 and master.

A proper fix would be to abstract the
view manager behavior as a mixin and use it for the selectcreatepopup
widget, for the one2many view manager and for the viewmanager.  But this
will take too much time for now, so a quick fix will do it. (however a
task has been created)