[FIX] sale_stock: Cancel all procurements linked to a sale order During the creation of a Sale Order containing a product for which 'Make To Order' is activated in the Procurements, two procurements are created: - one for 'Partner Locations/Customers' location - one for 'WH/Stock' location
If the Sale Order if cancelled, the user must first cancel the picking (delivery order).
This will trigger the cancellation of the procurement with the 'Partner Locations/Customers'
location.

The new behavior is to cancel the procurements linked to the Sale Order, which will trigger
the cancellation of the picking.

opw: 630170
issue: https://github.com/odoo/odoo/issues/3805