[FIX] kanban: wrong rotate of folded column in IE9
In IE9 the title of a folded column is rotated two times. The writing mode
"tb-rl" already rotate the bloc then the "-ms-transform:rotate(90d)" rotates it
a second time.

This fix set the writing mode to the initial IE9 value (with a CSS hack so
it doesn't impact other browser).