[FIX] stock_account: Create invoice journal type for transit.
This revision is related to 4fb497b65302a0350eda4f975627aa4e43ec9ef3.

For a 'To be invoiced' delivery order with a move with
as source location 'Stock' and
as destination location 'Transit'

There is no way to know if it's a Customer invoice
that should be created or a Purchase Invoice.
Indeed, the transit could be used as an intermediate to ship
to the supplier, or to the customer

opw-639536