[FIX] website: display 'Show Sign In' option
Commit 4ff1af4 moves the groups attributes on the ir ui view/template.
So the option is no more available in customize menu if user is logged.
No luck, because when your are not logged you don't have the menu.

Groups on template are the best practice to hide the content of a view
to a group of poeple, except in this case where a customize_show is set
to True.

Maybe that customize_show should ignore groups on ir ui view in a future
version ?