[FIX] tests: add missing icon
The menu root "base.menu_tests" is defined in 4 different places, so,
the presence of the icon depends on the order of installation.

This is a quick fix, it works because odoo serve static files of every
addons, even if not installed.  But this is code duplication that should
be removed in an ideal world.