[IMP] web_kanban: optimize drag and drop
This is a really difficult optimization.  Based on various benchmarks
and some hard thinking, we discovered that this change will shave
approximately 100ms for each drag and drop performed by a normal user.

We believe that this is a safe optimization in most cases (but we didn't
test it on some architectures, such as ARM or Solaris).  It is also
compatible with most SQL JSON-enabled databases.