[FIX] web_kanban: race when loading view_form
When the view form is loaded passing by the kanban list, the kanban
dataset data could be loaded after the view form is displayed.

When going manually to the view form, the human is usually enough slow
so the dataset ids are already loaded, but when loading directly the
view form (for example an "Edit" link from a record on the website or a
refresh) it can not be. For example on runbot the issue happened more or
less 50% of the time, and on localhost 90% of the time.

closes #7525
opw-642636