[FIX] web(_tests): FieldMany2One: autocomplete
Many2One fields were broken in views opened by default in edit mode, since rev.
a0ddaf8. The autocomplete widget was appended to the DOM at
'.oe_application > *:visible:last', but since the mentionned rev., the views
are rendered in fragments detached from the DOM, and once they are ready, the
old content is detached and the new one is attached instead. It means that by
appending the autocomplete widget within '.oe_application > *:visible:last',
it is detached from the DOM with the previous content, and Many2One fields can
thus not work.

This rev. simply removes the appendTo option of autocomplete, so that the
widget is directly appended to the body.

Also updates a tour's selector matching an autocomplete ul as it isn't inside
the .modal anymore.