[IMP] account: do not hide Journal and Period in Journal Items
The Journal field is not hidden anymore when filtering on the 'journal_id'. The
reason is that the filter might correspond to several journals. Therefore it is
not possible to know what is the journal corresponding to a given journal item.

The same is applied to the Period.

Fixes #7793
opw-646292