[FIX] models: access error due to prefetch of indirectly referenced records
Complements commit af9393d5059a41bc3bf7042ffdfefa070b7506c5
in light of commit 62b0d99cfe537e7f7f1e8794974230abbcdbfe44,
to really have the correct effect.

When the prefetching failed due to the presence of
extra records in the cache (for which the access is denied),
the `read` operation was indeed retried. However the
result was not stored in the cache because the cache
already held a FailedValue (automatically added when the
prefetch failed).