[IMP] base: config: module category cleaning
- removed unused categories, aka report designer, CRM and knowledge management
- renamed Warehouse to Inventory
- defined all user groups in base_security, instead of having most of them in
base_security and 3 others defined in other files. Portal, public and
partner_manager groups are now present in base_security. This uniformize
the declaration of groups and module categories.
- define a new Extra / Other category, holding those 3 groups, to ease the
classification and display
- rename Usability ot Extra Rights[IMP] base: config: module category cleaning

- removed unused categories, aka report designer, CRM and knowledge management
- renamed Warehouse to Inventory
- defined all user groups in base_security, instead of having most of them in
base_security and 3 others defined in other files. Portal, public and
partner_manager groups are now present in base_security. This uniformize
the declaration of groups and module categories.
- define a new Extra / Other category, holding those 3 groups, to ease the
classification and display
- rename Usability ot Extra Rights