[FIX] account: account type of account created on bank account creation
When creating a new bank account
e.g. Accounting > Configuration > Accounts >
Setup your Bank Accounts

When the user leaves the journal blank,
a journal, and an account associated to this
journal, are automatically created.

The account type of the account created could be wrong,
as it used the account type of the parent of the first
account of internal type `Liquidity`, which
could not be an account of account type Cash or Bank, but of
account type 'View', and such an account type does not
have the right delivery forward method, in order to report
correctly the amounts when closing a fiscal year.

Instead of using the account type of the parent,
it should actually uses the account type of the sibbling,
which have a correct delivery forward method

opw-647311