[FIX] website: condition to flag a menu as `active` in the top menu
1. A menu with `/page/website.***` should be flagged as `active`
if the current url is `/page/***`. This is a retro-compatibility
patch for c9d41679fbeebf244a75964f8875f9b6c6c9f4e1, so the
menu is marked as active without having to rename it, by
removing this `website.` thing.

2. If you defined two menus with as url `/page/test` and `/page/test2`
Both menus were flagged as `active` when you browsed the url
`/page/test2`, because it started by both menus urls.

Fixes #3059
Closes #3070