[FIX] account_analytic_plans: ACL for distribution plan instance
In the standard account_analytic_account module,
the employees have the read access to the analytic accounts.

Therefore, it makes sense to give the read access to the
account distribution plan, so an employee can choose one
in a purchase order when the Analytic Accounts for Purchases
is activated.

opw-647606