[IMP] crm, hr, project, sales_team: mail-related fixes
- hr, project, sales_team: no subscribe on create

For header records like sales steam, project, jobs and department, do not
subscribe the creator. Indeed those objects allow to automatically follow
child records (opportunities, tasks, issues, applicants). However those
records are often created by IT / technical / administrators that should
not be followers of all child records. Therefore the no subscribe key
is added in context.

- crm: new lead subtype

The 'new' stage is based on the sequence only, not the probability. Indeed
a new lead does not have a probability of 0; otherwise it is already lost.