[IMP] sales_team: invoiced are now a computed field
In sales_team, we had 2 fields for invoicing : invoiced_target and invoiced_forecast.
As forecast is useless because someone has to set it manually, we did something different.
The field invoiced_forecast is now invoiced and is computed according to the invoices
associated to the sales team. We can now only set the target.