[IMP] stock: pack operations read-only compute fields and product_id_change should set lots_visible
In other words, the 'split lot' button should be visible immediately
when choosing a product with lot/serial number tracking and this will
also make the quantity field read-only.

Compute fields do not need to be set as read-only.
It gave problems with the old onchange that needed to set it.