[FIX] google_calendar: force internal id only on creation
This rev. is related to 1c533b193feb7629d87ab1ff740b0cbc7615812e

The reason why we force the event Google internal id is
explained in the above commit.

Forcing the internal id must be done only when creating the
event in Google, not when updating, for retro-compatibility:
if the event is already in the attendee calendar, with a different
internal id then the other attendee, we must leave it like this,
otherwise Google will be lost.