[FIX] website_quote: Amount before discount with included taxes
The undiscounted amount did not take into account the taxes
included in the price. Included taxes were seen as a discount.

For a quotation with one line with an included tax of 15%,
but no discount, the undiscounted amount must be the line price subtotal,
the price subtotal already removing the taxes from the price unit.

For a quotation with one line with an included tax of 15%,
and a discount of 10%, the undiscounted amount must be the line price subtotal,
plus the amount discounted. That way, the taxes included are already removed
from the price, but the discount is not included.

opw-648388