[FIX] web_kanban: column & item quick-creation
In all browsers the ``focus/blur`` event is triggered before a button's
``click`` (it's actually triggered by mousedown).

Column and record quick-creates discard the form on blur but still want
to create the column/record when the "Add" button is clicked. This was
implementing with a fixed debounce/time-delay on the blur handler,
however the exact delay between blur and click dispatch are browser,
environment and possibly load-dependent (on my system unloaded, 10ms is
sufficient in webkit but firefox requires at least 50, a user keeping
the mouse button pressed for some time would increase that delay).

The proper way to handle this is to suppress the ``focus/blur`` event
from the relevant button entirely by preventing the browser's default
mousedown handler.

Do that for both column and record quick-creates, and remove
time-delay/debounces.

Also remove manual proxying of ``click`` in kanban column (move to
``events`` object)