[FIX] web_kanban: record padding on mobile
There are several problems with kanban record :
* Ideally they should not use row/col as the grid gutter is too big for
small cards like them
* Some template use rows but kanban record padding is not the same as
a container (and so rows are overflowing the record)

This does not induces a serious problem in desktop version because there
is enough space aroung the records to not have a visual problem.
However, this induces an horizontal scrollbar in mobile as the records
are full width in this mode.

Some solutions were possible :
* Change all record padding to be equal to the bootstrap container
padding -> not chosen because this would impact every record and the
gutter will be too big and make many records ugly
* Overriden inside-record rows and cols to have a smaller gutter
-> not chosen because this will add ugly rules alongside impacting
every record

The chosen solution is to increase the horizontal record padding in
mobile only. This will even align the record content with control
panel content.