[FIX] point_of_sale: fix 'Group of Taxes' calculation
There were still references in the POS to a child_ids field on
account.tax. This field was renamed in accounting v9 to
children_tax_ids. Because child_ids wasn't loaded anymore this meant
that child_ids on the taxes in the POS was always undefined. On top of
that the original code used Array.prototype.concat() assuming it
modified the array, when in fact it returns a new array.

All of this meant that group taxes in the POS were never calculated,
this resolves that.