[FIX] sale, sale_crm: convert opp to quote, do not force note
When converting an opportunity to a quotation,
the note was forced to the `company_id.sale_note`,
which could be `False`.

This prevented the other possible defaults for note to be set,
e.g. the note from the `website_quote` `template_id`.

Besides, the note forced in this wizard:
`sale_obj.get_salenote`
was almost the same than the defaults of the sale.order:
`company_id.sale_note`
except the language, the note translation in partner lang.

Therefore, the default value of `note` for sale.order has been
changed to use this `get_salenote` method, which handle
the language, so it's no longer required to pass it in
the convertion wizard, the note will now be added
as any default value, giving the possibility
to add another note default value.

opw-650050