[FIX] web: invisible element position in a listview editable
Elements which are not displayed (display: none) cannot be positionned
correctly by jQueryUI in a listview editable. Therefore, the solution is
to make them visible (but opacity: 0) during the time their position is
being computed.

This was a problem because a listview editable field may become visible
under some conditions AFTER their position was already computed.