[FIX] ir_qweb: assets: write in ir.attachment outside of main transaction
If, in the same transaction, we need to render a template and read its
associated assets, they are unreachable because the writes in
ir.attachment are not yet commited. Workaround this issue by commiting
manually with new cursors.

This issue occurs when you want to print a pdf report for the first
time.