[FIX] point_of_sale: set taxes when creating pos_order from form view
When creating pos_orders from the form view, tax_ids on the
pos_order_lines were never set.

This was kind of fixed in 5327869356ca629728ef45ebdef2598c1d02dd4b. But
in there it was in onchange_qty, because the way the fiscal position was
taken into account required it to be. Because we take the fiscal
position into account in a different way now we can assign the taxes in
onchange_product_id which makes a bit more sense.