[IMP] stock: deactivate picking types instead of using 2 security groups for not showing internal picking types in kanban
Instead of using 2 security groups to make sure
internal pickings are not shown in the dashboard
when not in multiple locations,
picking types are deactivated when the warehouse
is one step in/out and not in multiple locations.

That way it will work the same for other kanban views.

The colors of the kanban view are also changed.