[FIX] website_sale: ensure payment transaction is for current order
When ordering on the ecommerce,
if a payment transaction was found in the session,
this transaction was used as transaction for the current order.

Nevertheless, if the transction is no longer linked to
the current order, we should not use it.

This happened, for example, when the quotation
was deleted while the customer/user didn't close
its browser, and the transaction id was therefore
still in its session.

opw-650417