[FIX] stock: when transferring (action_done), do several links for the same operation and move at once
As the link between moves and operations is done for every quant,
we can do all links between a same operation and move at once.

Besides helping in performance, this fixes problems to handle all the lots
at once for one operation.

The views for pack operation lots are also refined.