[FIX] stock: allow moving a pack with a product with lots and main_domain and action_done should be rebuilt every time
Main_domain for getting preferred_quants needs to be reset every time.
(otherwise the method will change it every time and it will lead to inconsistent results)

An entire pack can be moved also when the products inside require lot tracking.  (check_tracking update)
There is no check if the quants have lots when a pack is moved entirely.
On the other hand,  the "put in pack" checks that what is put in it has the necessary lots.
(but if incoming is without lots and outgoing with, you can still do an entire pack out without lots)