[FIX] mail: fetch channel info on new message notification
When receiving a new message, we add it in the client side cache of the
corresponding channel. The problem was that when we previously unsubscribed
to this channel, it was removed from the cache, producing a traceback when
trying to add the new message.

This rev. checks whether or not the channel is in the cache, and if not,
performs an rpc to fetch its info before adding the message.