[FIX] hr,l10n_in_hr_payroll: ACL errors for users who are not HR Officer/Manager
These menus were visible to all employees, but the reports use
hr.salary.rule.category that is only available to HR Officers/Managers.

As these reports are not meant to be available for employees anyway, we
hide them.

(this bug has been spotted by the nightly crawler that clicks on every
menu as demo user)