[FIX] base: don't add last separator if no add attribute
When using "remove" in a xml inheritance without "add" attribute, the
separator would be erronously appended at the end of the modified
attribute value.

e.g: we have "a,b" from which we remove "c", the end result would be
"a,b," which is an invalid jQuery selector.

This commit change this so empty tokens are not added to the attribute.