[FIX] mail: disable message_last_post to prevent race conditions
The message_last_post mechanism was introduced to have a trigger for
automated actions, but for some reason we do not use it.  Moreover, it
interacts badly with the bus.

Because of the message_last_post update, concurrent transactions can
fail and be rollbacked.  This is an issue because the bus commit the
transaction after writing the messages.  This means that some messages
can be duplicated, and the more users are chatting, the more frequent
this will happen.

Since we don't use this field, it could be removed from the models, but
we are in a stable release, so we need to keep it for now.