[FIX] web_editor: ie issue with range within undisplayed
In IE11, when adding a range to a selection that can fail if this range
is not displayed (display non on itself or an ancestor) or is not already
inserted in the dom.

e.g: IE fails in the following situation: http://jsfiddle.net/7cmf2tt9/7/

The feature causing this situation is minimal : it only set the cursor
at the end of a FieldTextHtmlSimple widget (the default widget for
fields.Html). This commit thus only ignore that error.

note: this is not an issue on IE Edge.

closes #9216
opw-652347