[FIX] web{,_kanban}: too much and too little shortcuts
When several accesskey with the same key,  are defined on elements, the
last one (even if not displayed) is the one which will be actioned and
the key sequence is pressed.

This caused an issue since the "Pager" template is shared between all
elements, even the x2many relational widgets.

Also there was shortcut for Create/Save/Discard on the form view, but
those shortcuts should also be enabled in the list and kanban view.

closes #9267
opw-652888