[FIX] bus: share longpolling among tabs
The bus used to do a long poll as soon as it is started.  This is fine
when the user has a single tab, but not so fine when 5 or 6 tabs are
open.  This commit introduce the concept of tab manager, a way to decide
among all tabs who is the leader.  The leader will then initiate the
long poll, and share the notifications via the local storage.