[FIX] base: avoid unlinking demo partners in tests
Unlinking a partner will also delete the related attachments (including
binary fields ones). When we rollback the transation at the end of the
test, the deleted attachments are not recreated in the filestore. This
unfortunate side-effect cause traceback when accessing these records in
backend.
In order to avoid this, we create disposable partners to test the api
cache behavior.