[FIX] sale, website_sale: send email at sale order confirmation in ecommerce
When confirming a sale order through the ecommerce
(at payment, most of the time),
the confirmation email should be sent at confirmation.

This is the subject of the revision
809e1a60813e0468ffc3f52f866119bb8cfe3a58
that was lost with revision
94716a3f14d929574c46ef165e4364d12a5903ed

At the moment, the email was sent only for
orders not being automatically confirmed
(transfers, for instance).

opw-653800