[FIX] mail: issue when leaving channel and refreshing
When we refresh, the bus will send the last notifications since the last
50s, so, the browser will receive notifications from a channel that we
already left.  This commits tries to fix the issue by sending
explicitely the information that a channel was unsubscribed.

Also, and this is touchy, it ignores messages received in a channel that
it is not subscribed to (before, it joined the channel)

Also, it will synchronize channels opened and closed in various
tabs/browsers.