[FIX] web_planner: hide background pn planner destroy
A PlannerDialog does not extend Dialog for historical and graphical
reasons. Thus sometime a logic in Dialog is missed in a PlannerDialog
whilst it is useful.

In this instance, when doing an action from the planner which destroyed
it, the modal backdrop was not removed by bootstrap. Bootstrap does this
when the modal 'hide' function is called.

opw-653734