[IMP] base_action_rule: manage recursive triggering of automatic actions
The base action rules manage a kind of checklist with the records already
processed for each action.  This checklist may be populated by recursive calls
on other models (such as the update of relational fields), and guarantees that
a given action on a given record is executed at most once.