[FIX] mail: needactions: mark as (un)read
Marking a message as read in the inbox doesn't remove needactions on all
channels anymore.

When joining a channel (other than the inbox) in which there are needactions,
those needactions are automatically marked as read and removed from the inbox.
When the user clicked on Undo in the snackbar, the messages weren't correctly
re-inserted in the inbox (in the JS cache).