[FIX] web: editable list validation with several pages
This revision is related to a9d20d9fa04b1dbb350663652e0996e258ac6d31.

`dataset.cache` contains an item for each records, even
if the records hasn't been loaded in the cache. If a record
has not been loaded, only the record id is available, not
the other values. Thefore, this isn't possible to validate
the record values of records that have not been loaded
in the dataset cache.

opw-654013