[FIX] im_livechat: missing dependencies in im_livechat page
The Website Live Chat is supposed to appear following 3 ways outside of
the backend:

- in a singleton page with path /im_livechat/support/{id}

- in a page in which a code code snippet (with two script and a css
  bundle) has been inserted in the head section

- by installing the module "Website Live Support"

In the first both instances, including some less files and bootstrap
default variables is required which this commit do.

In the last instance these file should already be included in the
website and the additional assets required are inserted in the
website.assets_frontend bundle.

opw-654898