[FIX] product, sale*: procurement needed
In v8.0, a procurement is not created by default. Only sale_stock and
sale_service will create procurements, since they are only used for
physical products or task-related services.

In v9.0, the behavior was changed and a procurement was always created.
This is not necessary, since the procurements are still only used for
physical products and task-related services. This behavior will just
lead to the creation of procurements which will fall into exception
(method `_assign` will return False).

This fix switches back to the previous behavior. The other solution was
to prevent the exception, but it will imply that useless procurements
will pile up.

opw-654363