[FIX] project: project and stage views
project.project: alias should be visible without being in technical group.
Indeed the alias is displayed in the kanban view. However in order to edit
it the user has to go into technical mode, which is confusing. The alias
is now displayed in the form view, but advanced alias configuration is still
in technical group.

project.stage: show projects using the stage. Without that field it is quite
impossible to configure the use of stages in the project application.