[FIX] stock_dropshipping: multicompany filter
The multicompany rule on procurement.rule may not be working as the field
company_id is not displayed with no default value (so probably often empty).
This means that the user may have access to rules from another company from the
product but failing to read the stock.picking.type

To avoid regressions introduced by 35f3dea, make a search on stock.picking.type
to filter on the one the user really have access to.

opw 658141