[CLEAN] project: dead / unnecessary code
Removed unnecessary imports
Removed duplicated _order
Cleaned kanban state update in write to avoid useless loops. Indeed when
updating the stage the kanban state is reset, if the new stage is different
from the old stage. Now in every case the kanban state is reset. This is
already the case in issue and does not seem to frustrate users.