[FIX] point_of_sale: make tendered amount adhere to currency formatting
Part of the number formatting overhaul done in
5c9b4556e35c420985696b0c958ca492030efa4a,
4b003afd0ab125b0296de8939626ae889ccdab47,
261ccf7c56062930b1d712ae860fe01dd8718968 and
6e945de681452d6753f58a749352bfbe099bac1d. The only place left where we
didn't correctly format the numbers was in the tendered column on
the payment screen. This fixes that by not directly displaying the
inputbuffer anymore. Instead we pass the payment line amount through
format_currency_no_symbol() and display that instead. We also give a
default amount of 0 to parse_value so it doesn't return false for an
empty string.