[FIX] point_of_sale: subtotal non ESC/POS receipt shouldn't include tax
On the PosTicket receipt (non-ESC/POS receipt) we have a line titled
subtotal. The subtotal was calculated with getSubtotal(). The issue is
that getSubtotal() ends up just using the price of the product. This
doesn't make sense when taxes are included in the price, because then
the subtotal will contain taxes.

It's not an issue on the ESC/POS receipt, because there we don't show
the subtotal at all if we only have included taxes.

fixes #10137
opw-660266