[FIX] sale: A new 'task' line should not generate (and overwrite) the sales order analytic account
(A or B and not C) => (A or (B and not C)) which means that
A will always cause the creation of a new analytic account
regardless of C (which is checking if the sales order already
has a 'project_id').

The 'poject_id' check should always prevent the call
to `_create_analytic_account()`.