[FIX] stock: consistent label of field
Field virtual_available on product.product is called Forecast Quantity but was
not on product.template.

This is not an issue to modify in stable as the term is already translated and
the translation of the string of the field is not based on the source (meaning
that, without reloading the translations, people will still see the translation
for 'Quantity Available').

Closes #10443