[FIX] product_visible_discount: trigger property on overriden `product_uom_change`
This revision is related to feb2ede0d6c48982209a5b18a20b58f69c70327c

When overriding an onchange method, the `@api.change` property,
that decides on which fields the onchange method must be triggered,
should be the same on the overriding and the overriden method.

Otherwise, the base onchange won't be triggered on the fields
it expected to.

opw-668049