[FIX] account_budget: To print the summary of a budget
The theoritical amount is computed as:

theo amt = ellapsed time / total time period  * planned amount

where:

total time period is the difference between date_to and date_from defined
on the crossevered budget line.

ellapsed time is the difference between today and date_from.

The achievment(percentage) on a crossevered budget line is computed as:

percentage = practical amt / theo amt

where:

practical amt is the sum of all account analytic lines valid in the
crossevered budget line period.

When computing the report of a summary, the ellapsed time is equal
to the difference between date_to and date_from defined for the summary and
the practical amount only takes the account analytic lines in the summary
period.

opw:669545