[FIX] sale_margin, delivery: margin at creation
The `purchase_price` is filled in automatically thanks to the onchange
method `product_id_change_margin`. However, when a SO is created
programatically, it is necessary to call manually the method in order to
fill in the field. This is not convenient since the module sale_margin
might not be installed, and crappy workaround (such as the use of
`hasattr`) are necessary.

To avoid this, this fix automatically computes the purchase price at
creation if no value is provided.