[REF] web: view_manager's init
Always initialize a view_manager with its dataset and views, but without the
action (at least explicitely, as the action might still be given within the
options).

This is a backport of changes in enterprise to keep the view_manager's API
similar in both repository, which is required the large refactoring that will
arrive.