[FIX] sale: mail template
The URL sent to the partner in order to view the quotation is not
usable: the report/pdf route is not public, and therefore a partner
without an account is not allowed to download the PDF report.

The fix is to sent a sign-up URL instead, as it used to be. Note that
the portal functionality should be removed in the near future, and
another fix will have to be found.

opw-667606