[FIX] product: do not add same attribute on product
Adding an attribute with only one value does not recreate the product,
but write this attribute on every product instead. However, the current
filter is not correct: it checks if the given attribute value is not
already in the list of attribute values of the product.

The appropriate check is to verify that the attribute itself is not
already in the list of attributes of the product. Indeed, there is a
constraint to prevent the same attribute to have two different values on
the same product.

BTW, it's a new api function (`mapped`) inside an old api module. Not
the cleanest way to do it, but simpler to write.

opw-671114