[FIX] barcodes: first qualify the quantity event before handling it
Introduced by 5bb64d40d0b94b252c2e29e6a050ae37fbd87b0f. Before doing
anything we should first ensure that the quantity keypress event targets
body (and that it is a number, but that's less important). Before this
we didn't do that and thus the _display_no_edit_mode_warning was given
when people where just typing in the associated thread of a picking.

opw-672350