[FIX] point_of_sale: correct sign of fixed tax
Oversight of e8b965f93c3eeb541aba7b6a435193a9008516ac. Although the
previous commit gave the correct behaviour for a negative fixed tax with
a base_amount of 0, it was still incorrect for positive fixed taxes with
a base_amount of 0.

This rewrites the code to more closely resemble the relevant code of
account.tax._compute_amount() solving the above mentioned issue.