[FIX] pricelist: no early rounding of computed price
The pricelist computation should not immediately
apply the currency rounding to computed prices,
to allow for unit prices with arbitrary precisions.

This way users can use a higher decimal precision
for "Product Price" than what is allowed for their
currency. This is useful e.g. to obtain exact
subtotals prices when combining small prices with
large quantities.

The resulting price will be rounded appropriately
when it gets stored to the destination field anyway.