[FIX] project_issue: link to issues
On the project dashboard, a user needs to click exactly on the number or
on the label in order to open the issue list. If the user clicks outside
the text (but still inside the box), the project is open instead.

The reason is because the link is specifically defined on the
number/label. The fix defines the link on the box itself.

Fixes #11396
opw-672380