[FIX] mail: translate static channels
The name of the static channels ('Inbox', 'Starred') are not translated
in the breadcrumb. The reason is that at intialization of the chat
manager, the translation are not yet loaded.

The fix is to use the lazy translate functionality, which applies the
translation when the string is displayed.

Courtesy of @aab-odoo

opw-673856