[FIX] mrp_repair: tax computation
This solved two issues. First, the price subtotal should contain the
price with taxes excluded, as it is the case in other modules.
Otherwise, the tax will be computed twice. Then, when an invoice is
generated, the taxes should be calculated...

This is an oversight of commit fe2b5f35a4.

opw-673984