[FIX] account, sale: modify invoice
When a user validates an invoice linked to a SO, he might want to modify
it afterwards thanks to the 'Modify' option in the refund wizard. This
will create a refund, reconcile and create a draft invoice.

The issue is that the draft invoice is not linked with the SO.

The fix is to link, in this specific case only, the draft invoice to the
SO. At the same time, we have to make sure to:
- unlink the first invoice (to prevent counting the quantity as invoiced
  in the SO)
- keep track of the first invoice/refund in the invoice list of the SO

opw-673446