[FIX] account_asset: deferred_revenue_category_id and asset_category_id
When creating an customer invoice with a product linked to a deferred revenue
category, the account used in the invoice line must be the account_asset_id.

When creating an vendor bill with a product linked to an asset
category, the account used in the invoice line must be the account_depreciation_id.

Inspired from 0287c1e1b34a14300dcf7ef9ffddf160b61631ee

opw:677098