[IMP] product: remove unused field `name_template`
The field `name_template` is removed. There is no real use of this field
in the Community or Enterprise versions.

Moreover, an issue occurs with it: it takes the value of product
template name, whatever the language of the user updating the record.
In other words, if a user changes the name of the product template in
order to define a translation, this translation will be used in
`name_template`. This could be solved by patching the ORM
(http://pastesha.re/ihRc), but patching the ORM for a single field with
no real use is not worth the risk.

opw-678143