[FIX] web_settings_dashboard: Incorrected plural form of "Active Users"
Odoo add the extra letter "s" to the translated word "Active User".
This is caused by adding "s" at the end of the word directly from the template,
which makes no sense in many language.
This prohibits translators from translating the plural form in any language.
The plural form and singular form of "Active User" should be translate
separately.

Fixes #12469
Closes #12470