[FIX] base: add disabled users at write on 'users' or 'implied_ids' of res_group
If a user is disabled but in a implied_ids group, it was ignored and
so not added.

Eg: public_user is by default disabled,
So:
    <record model="res.groups" id="base.group_public">
        <field name="implied_ids" eval="[(4, ref('foo'))]"/>
    </record>

will not add the public_user in group foo.

It is coherent with the method execute of resConfig which already use the
active_test=False.

Courtesy of @odony for review