[FIX] web: fix label for selection fields in grouped list view
Before this commit, the list view, grouped by a selection field, displayed the
'technical' value of the field as group label.  It was basically forced
to do that, because the result of the read_group only contains the
technical value.

This commit changes the list view in a way that it now requires a full field
get, so it can use it to find the correct string.