[FIX] stock: receive product with lot
When a product is received in the lot reception view, a new lot is
created. If the user receives at least one lot, closes the lot selection
window, then reopens the window, he will not be able to create a lot
anymore, but only choose existing lot.

The fix is to make sure that the lot selection is available only in the
case of a returned product.

opw-681925