[FIX] account: bank statement reconciliation speed improvement.
Removed assertion statement that was making a useless browse on account.move. This single patch is enough to divide by 4 the time of preprocessing a bank statement when opening its reconciliation widget.

That removed assertion is now a supposed prerequisite when calling the function and should be tested first (on batch of records).