[FIX] account: unnecessary default currency
A mechanism already exists to select the appropriate currency and was
introduced in commit 1b04b1d. However, it requires a purchase journal in
the currency of the PO to work.

Note that this commit is likely to not work correctly since the method
`_onchange_journal_id` is triggered after `_default_currency`.
Therefore, even if the currency is correctly set in `_default_currency`,
it is changed to the journal currency right after.

This reverts commit 7f1d7c1cce6dac89695c57d8507b609aecaf0c51.