[FIX] sale_timesheet: don't count expenses as timesheets
On sale orders, the timesheets stat button should display the number of related timesheets.
0b55f2ee6d55365ae0f06473585d543f1cc7b259 changed the computation system to have the number of timesheets
instead of the total time for several reasons, but used a domain which also includes expenses.

This commit changes the domain to make sure that the stat button only displays the number of timesheets,
and that the action leads to a list view of timesheets only.