[FIX] point_of_sale: total revenue in report
In some cases, the "Sales total(Revenue)" of the report is different
from the sum of the payments. To reproduce:
- create a product with a sale price of 1.00
- sell 3 times a quantity of 0.333 of this product
- the "Sales total(Revenue)" will be equal to 1.00, while the total of
  payments is 0.99.

This is because in the report, we first sum all order lines, then we
perform a global rounding.

The solution is to use the total amount calculated for each order, which
is rounded at the appropriate time.

opw-682782