[FIX] account_budget: reset correctly result variable
If a budget line has no analytic account id, the code is intentional
about returning a value of 0 for it.  It is possible, that in some
cases, depending on the order of lines processed, the routine returns
the previous value, as it does not initialise an important
variable on each iteration of the loop.

Closes #13303