[IMP] mrp: routing of phantom boms in bom
When the routing of the phantom bom is the same as the parent BoM,
the phantom BoM should not recreate the work orders, but use the same
work orders as the parent BoM.  That way, you can create one master
routing and the "consumed in operation" of the phantom bom can be used
linking to the master routing.

If the phantom bom has no routing or the same, it can also use
the "consumed in operation" of his parent bom line if none is defined.