[FIX] website: sign_up page resize
The less code that was resizing the login and sign up form container
in v9 was moved to the login form, but not to the sign up form.
Adding the sign up form class to the less code fixed the size.

It has been done in rev https://github.com/odoo/odoo/commit/5b65befcd889139152c590490fb8965c9f4b5279