[IMP] mrp: active_move_lot_ids needed on stock_move as well, not just on workorder
As you open the lots in a raw material move in a production order,
you should not see the temporary stock move lots from the work order.
Also the quantity_done should not take into account those.  They will
when they press 'Done' on the workorder.

When posting inventory, the temporary stock move lots on the workorder,
should be transferred to the split move.

For splitting and unreserving, some optimizations are done.