[IMP] product, website_sale: add field sequence on pricelist.
Now that property is computed automatically.
In some case, the compute will take the first on available.
So, the order (sequence) allow to specify the priority.

Test from point of sale are base on currency $,
So we force for the test to use a pricelist in dollard (My company, Chicago)