[FIX] web_editor: stop resetting bg image of parent snippet on color change
Commit 404ed0f7884ff724706e6d7f50d35482e92ce60a tried to solve the
problem but only solved a part of it. Indeed, if the snippet child
has a bg color option but not a bg image option, the event propagation
stops in the parent handler, when it is too late.

Solution is checking that the event comes from the same element on
which the callback is attached to.