[FIX] report: Allow a portal user to print the 'external_layout_footer'
A user portal cannot have the read access right on the res partner, except his commercial_partner_id (his partner or the root partner if he got a parent company). That would a big security flaw.

In the header we use the contact widget to print the company information. This widget uses sudo to get all the required fields to display.

But in the footer, we still access some fields like company.fax which is not allowed.

As the contact widget uses sudo to display the company information in the header, we did the same here for the footer.