[REF] sales_team, web: remove sales_team_dashboard view
The view `sale_team_dashboard` (which extends kanban) has been removed
and replaced by a default kanban view.

To use it in the ViewManager, an option `js_class` is added
on the arch description and this class be used as the View.

This avoids creating a new view type just for one view.