[FIX] web_editor: allow to remove image from gallery thanks to remove button
Commit b309f3ab543b5195eedc5b06ad08cfa81f43171a made the snippet remove
button always be there by removing the "preventParentEmpty" mixin.
Instead of preventing to remove the only child of a snippet, the system
was changed to remove the parent if this only child is removed. This
also applies on the system which hide the remove button if the child
cannot be dropped anywhere in the DOM.

Unfortunatly the image gallery snippet relied on the remove button to
be hidden for its content images. Indeed, it was implemented so that
the "remove" button is hidden but a "remove image from gallery" was
there in the customize menu. The solution is to remove this option and
copy its behavior on the standard remove button.

Note: this probably was not done before because the on_remove feature
was broken (fixed with commit 5e9ebcfb01f3250833a543154d072d6c785c9330).