[FIX] website_crm_partner_assign: fix portal lead/issue to use commercial_partner_id
Before this commit, only partner really assigned to the lead AND with a grade was
able to see their leads.
Now we check that the top level partner (commercial_partner_id) has a grade, and show
all the leads under this commercial_partner_id.

Try to get the same behaviour that previous portal backend.

opw-688359