[IMP] ir_sequence: ease debug in multicompany dbs
It often occurs that users get nasty error messages like "missing value
for column xxx" when trying to create objects. It can happen if the
objects uses a sequence as default value that is not defined in an
available company.

We try to give a hint to the glory support person who could be confused
by that.