[ADD] point_of_sale: differentiate the sale and invoice journal
Using the same journal for both the session closing and the invoices generated
from the PoS may be problematic as they share the same sequence.
In such case, gaps may appears in the invoice numbers (as the sequence is on the
account.move, closing a session increases the sequence by one).

cf opw-685660
Fixes 13242