[FIX] website_quote: /quote/ redirects to login when not signed in and no token
When the route /quote/<order_id>
is accessed without token
e.g. through the `View sales Order` button
in the headers of thread mails

the user must be redirected to the login screen
if he's not signed in.

In other words, you must be signed in to access
this route without token.

opw-687931