[FIX] rating_*: domain on m2o for rating mail template
Rating is implemented for project.issue, but it is
impossible to set a template on the stage since the
domain force the mail.template to be declared on
project.task.

We use the same mecanism as `mail_template_id`
declared on project.stage.task in project
module. The list of available template on the stage
form view will be the ones for issue and task like
mail_template_id.