[FIX] website_quote: correct company for the quote validation
In the frontend customer quote validation,
the company used in the sentence
```
I agree that by signing this proposal,
I accept it on the behalf of
```

was the quotation company, which is the company
which is selling, not the company/customer which is buying,
and this does not make any sense.

Instead, it should use the company of the customer
accepting the sales order, or simply the customer
if he does not have a company set.

The only case which make it a little weird is when
it's the customer himself, without parent company,
who is accepting the quote
"on the behalf of himself", but:
 - This isn't possible to correct this without
   breaking the translations, because the
   sentence "I accept it on the behalf on"
   is within the same sentence than
   "I agree that by signing this proposal"
 - This makes sense if another person than the
   customer himself is validating the order
   (the customer sent the quote link with the token
    to another person).

opw-690837