[FIX] account: tax base computation
Before the fix, the tax base amount is computed only for a complete
invoice line. However, depending on the tax sequence and their inclusion
in the base amount, the tax base amount might be different for the
various taxes of an invoice line.

We keep track of the base amount for every single tax, and refactor the
`_compute_base_amount` to take these values into account.

opw-691177
Closes #13819