[FIX] stock_account: valuation and costing method
In v9, the creation of a product doesn't set a `valuation` or a
`cost_method` by default. It means that the corresponding fields of the
product category is selected instead.

In v10, this is the other way around: a default value is set on these
fields, and therefore the value on the product is selected. This is
obviously a change in behavior, and moreover the field is invisible in
the product view, meaning that it cannot be changed.

This is due to the new API conversion. In the new API, the methods
`_compute_valuation_type` and `_compute_cost_method` are called through
the 'onchange' mechanism. This sets the associated fields, which are
then copied in their `property_` counterpart thanks to the inverse
function.

opw-691231