[IMP] sale_stock: sale order line quantity warning
When you add a sale order line, odoo checks the available quantity
through all warehouse and if you add more than available, you get a
warning.

We improved this warning, by checking the quantity in the warehouse set
on the sale order, and if the quantity available in this warehouse is
too low, we warn the user with the quantity available in this warehouse
and in all other warehouses.