[IMP] web: disable selected fields in pivot dropdown
In the pivot view, when the user select a new field to drill down, the
previous dropdown menu displayed all the groupable fields as valid
choice.  This commit disables all the fields that have already been
selected.  It means that they will be greyed out, and inactive in the
UI.