Revert "[IMP] website_sale_delivery: Allow to (un)publish delivery methods from website"
This reverts commit 6d9c59f348453a2c9026c4d9a5bb121790a7e189.

This commit is breaking the enterprise/master branch because the method _get_delivery_methods takes 5 seconds to be executed with some delivery methods. Now, instead of having 1 delivery method displayed, the admin has all the delivery methods in order to be able to publish/unpublish them. As this method is called for each carrier, the phatom js tour takes more than 15 seconds to display the page /shop/confirm_order in the ecommerce, with is more than the timeout threshold.

I revert this commit while waiting to find a proper solution.