[FIX] stock: fix stock picking done quantity computation
This is an oversight during the port of
3de53fa165baf76cde7f2593d4fb26f2cc7a5383
in the forward port revision
3a2147d95c8005ab7c1dbe4d94315e404ea5e5f9

In the revision 3de53fa165baf76cde7f2593d4fb26f2cc7a5383,
the code has not only been moved,
but the variable passed to the method `rereserve_quants`
argument `move_ids` changed as well.
This has been missed during the port of the revision.

This leads to the invalid computation of the `done` quantity
in the pickings in some cases were the quants need to be
re-reserved.

opw-690244