[FIX]account: customer invoices with analytic account
-amount_currency in model account.analytic.line is a related field to move_id.amount_currency
where move_id is an account.move.line

-amount_currency is computed in function create on model account.move.line by considering that
amount = debit - credit (since 8.0)

-amount in model account.analytic.line is computed by considering that
amount = credit - debit (since 8.0) in the function _prepare_analytic_line.

So when creating a customer invoice with an analytic account in another currency, the created
analytic line has an amount which is the opposite of its currency_amount.

This commit also reverts the commit 7fca8ad89eeb53844724c429e983085f8ddc6c4a to keep the same
behavior as in 8.0 for the creation of an analytic line from:

- an expense
- an invoice
- a task

The computed not stored field analytic_amount_currency displays the currency amount with the right sign.

opw:694085