[FIX] product: fix company_id assignment on pricelist version & items creation
Due to the following revision:
295b96c0b39959d053c3b74edd515879c873c722

The company on the version and items was not assigned
when creating a new pricelist version on an existing
pricelist.

They were if you changed the company of the pricelist,
thanks to the triggers added in the above revision,
but not when they have just been created in the pricelist.

These new triggers repairs this regression. It also handle
changing the `pricelist_id` of a version (this is doable through
the "Sales > Configuration > Pricelists > Pricelist Versions" menu).

Fixes #14631