[FIX] sale: fix company_id assignment on order line creation
Due to the following revision:
456d7b38f1f6c889ab24e391154dbe1d21f34371

The company on the order line was not assigned
when creating a new line in an existing sale order.

The company was correctly assigned when you created
the order line with the order, and when changing the company
of the order, but not when adding a new line on an existing order.

This new trigger repairs this regression.