[FIX] website_sale: Rename warning message for eBay settings
On the website settings, it was intended to have a link to the Sales Settings when sale_ebay is installed.

As there's no link between the website and sale_ebay modules and we didn't want to create a bridge module only for that, we decided to remove the link.

The warning in not coherent anymore and should be changed