[FIX] account_asset: Wrong account set in Vendor Bill
This reverts commit 804ef6f.

1. Generating the vendor bill by adding the related Purchase Order:
in that case, it works well since the system has a first on-change on the product when selecting the asset type;
it sets the asset account in the expense account of the product which in turn is used as the vendor bill account.

2. Creating the vendor by adding manually the product which is an asset:
in that case it takes the wrong account since it used the depreciation account instead of the asset one.

But the labels cannot be changed because the case 1 works.

opw:702265