[FIX] account: ir.mode.access on products
When the 'Invoicing' or 'Accounting' module is installed without the
module 'Sales', only the admin user can create products in the database.

The expected behavior is that any 'Accounting Adviser' should be able to
manage some products in Odoo.

To allow 'Accounting Advisers' to manage products, we add them the CRUD
rights on 'product.product', 'product.template' and
'product.price.history'

opw-703610