[FIX] website_sale*: default company & warehouse for website orders
By default, the default warehouse of a sale order is the first warehouse
of the company of the user creating the sale order.

In the case of a sale order created from the website,
the user who is creating the sale order is the superuser,
a.k.a the administrator.

There was a big issue in the context of a multicompany database,
with an administrator changing of company:
the sale orders coming from the website would suddenly change
of company, which is wrong.

This was even more severe when the admin changed to a company
which had no warehouse defined:
It was then impossible to create the website cart, as no default warehouse
was found, while the warehouse is a required field on the order.

This revision ensures that the cart / sale order is always created
under the website company

opw-704557