[FIX] point_of_sale: export orders and traceback now work for debugging.
In debug mode the buttons "export unpaid orders" and "export paid orders" now
work. On first click, they prepare the file and display the download link. On
second click, the file is downloaded.

The error popup uses the same method to download the traceback.
Use a different icon to have a visual feedback why that 2 clicks are needed.

This 2 steps is required as it is not possible in most browsers to simulate a
click event on a link.

The download method is kept for backward compatibility but will be droped in
master.

In the next version, a proper popup will be integrated (cf #14365)

Closes #15322