[FIX] sale, website_sale: enterprise symbol in settings
In the sale settings view, the field company research hasn't
the enterprise symbol though the related module is in enterprise.
It's the same problem for the Company Research option in the website
admin settings.

Both field declaration use a string attribute in the label tag.
but it is needed to use a "for" attribute. The Upgrade widgets only
put the "enterprise" span after the enterprise label, wich is found
by looking on node having a for attribute with "module_ " and its value.

This commit use an attribute "for" instead of the attribute "string"
in the label declaration.