[FIX] account_voucher: use currency rate at date for all amls
Without this the first account.move.line will be generated with the
currency rate valid at the time of validation. The following lines
however will be generated with the currency rate valid at 'date' of the
voucher. If these currency rates are different this results in an
unbalanced journal entry.

This ensures the first account.move.line will also use the currency rate
at 'date' of the voucher.