[IMP] stock_account: remove security group inventory valuation
It will remove the stock setting, the group itself and remove it where
it is used in the views.  The perpetual valuation setting is only on the
internal category which is already "deeper" in the system.

It is also removed for the update price link on the product form
in case of perpetual valuation and we changed the label to Update Cost
instead of update standard/average cost.