[FIX] point_of_sale: IE close itself on the POS opening
When a session was opened in Internet Explorer (11 at least), it was directly
closed.
This is due to the concurrency mechanism created to avoid having two tabs opened
with the same window.

To avoid having two tabs opened on the same session, the browser listen to
'storage' event, triggered when the localstorage is updated.
On most browsers, this is only triggers when the localstorage is updated in a
different window of the same browser.
On IE 11, it is also triggered when the localstorage is modified in the current
window.
It is unclear if this is a bad interpretation of the specification or not.

A bug has been reported[1] to Microsoft about it but is, as of today, postponed
without action.

To avoid the above issue, add a parameter containing a timestamp at the session
launch and be able to detect that the storage events comes from the current
window.

opw-724851

[1] https://connect.microsoft.com/IE/feedback/details/774798/localstorage-event-fired-in-source-window