[FIX] web: check if changes can be discarded before leaving view
Before this commit, we did not check if the changes can be discarded
when opening the app switcher.  This is now done.  Note that this check
was formerly done in the form view, and we moved this to the view
manager.  This now means that all views have the power to decide if they
can be discarded.