[IMP] im_livechat: rating improvements
This commit display rating on kanban card
of livechat channel, in order to directly
jump on the kanban rating used to compute
the customer satisfaction on the last 7 days.
Default filters are used.

A reporting menu entry is also added to have
access to global livechat satifaction reporting
with graph and pivot views.