[FIX] website_sale: product detail scroll
On mobile view, in the product detail page, the image of the product
fills the largest part of the screen. However, when the user tries to
scroll by clicking on the image, it zooms on the image instead of
actually scrolling. This is confusing for the users who report that they
cannot buy anything on the webshop.

A solution is to simply deactivate the zoom on mobile.

opw-704026