[FIX] web: deal wih `default_order` attribute
One could set the attribute `default_order` in the view arch
(like 'name,id desc'). The records should then be sorted accordingly.

This feature was not implemented in the new views.

This attribute should work on both basic views and embedded views (x2many).
The former uses the `sort` argument in `search_read` while the latter is
sorted client-side.

There is however a potential issue with the latter: as the order is done
client-side with records received by the server, if there are more records
than the limit, the server will send records without any order. The client
side order will thus be done on a fraction of the records.