[FIX] account: invoiced amount with down payment
- Create a SO of 1000
- Create a down payment of 300, validate
- Create the final invoice

In the partner view, a total of 1600 is displayed as invoiced.

In the case of a down payment, the corresponding invoice line has an
amount with a negative sign. Since the report computes the absolute
value before multiplying by the sign, this information is lost.

The absolute value is not necessary since `price_subtotal_signed`
already includes the correct sign.

opw-741159