[FIX] web: fix toolbar test (should not depend on document)
Unit tests are executed after module installation, which is sometimes
challenging when some part of the UI is modified by another module.
In this case, a test in module web/ was implicitely relying on a behaviour
introduced by the module document.
So, the test failed when document was not installed.

Since the installation of ´document´ module changes the toolbar display,
we check in the test if the module is installed to do the correct assertion.