[FIX] stock: PO line with unit_price = 0 and cost_method = 'real' or 'average'
Let's consider the stockable product P with an internal category set
with cost_method = 'real' and 'average'. When creating a PO with P
and a unit_price=0, after delivering the product, the cost transfered to
the stock valuation account  was equal to the product cost multplied by
the product quantity instead of the unit_price set on the PO multiplied
by the quantity.

PS: The quant is created with the stock.move in _quant_create_from_move and the unit_price
set on the stock.move is taken with function get_price_unit.

opw:740865