[IMP] mail: better navigation when expanding chat windows
Before this commit, expanding a chat window would open the discuss application, which means that the current navigation state (breadcrumbs) is lost.  With this commit, it simply opens a new action in the breadcrumb, and does not lose the current action. (PR 16725, closes task 31767)

Note that this implies that it is possible to have two discuss client actions open at once.