[FIX] crm: quick create in crm.lead
When making a quick create for a crm.lead record, the field user_id is set
with its default value (introduced by aa44700dccdc2618e0b8bc94252789264104047c)
But the date_open('Assigned') was not set.

When creating a challenge with goal="New opportunities", the new opportinities
are counted with date_open. But when creating an opportunity with the quick create,
it was not counted as date_open was empty.

opw:742751