[FIX] web: properly react when save operation fails (form)
Before this commit, the save method in the basic model did not properly
react when the rpc fails.  In that case, the returned deferred stayed
indefinitely in the 'pending' state.  This is a big problem, because the
basic controller put that deferred in a mutex.  As a result, every
subsequent operations (including field changes!) were indefinitely delayed.

To reproduce the issue, you can find any form view where some python
business logic will cause a fail in some cases.  For example, create a
new inventory adjustement, click on 'Start inventory', then add a line
in the x2many with the same product as another line.  Saving that record
will fail, but then, the form view is broken: no interaction will work.