[FIX] product_margin: graph view only use one non-stored field as measure fixes #16565
Because read_group can not handle non-stored fields well,
we fall back to the similar implementation as for the stock
valuation report.  This allows only one non-stored field as a measure in
the graph.  Then it is more logical to show the total margin as that is
what the module is meant for.