[FIX] web: SearchView: add custom filter on datetime field
Function 'set_value' of datepicker had been renamed to 'setValue'
by commit 5f13b7f. There were still a call to that function that
needed to be adapted in the search filters, so it crashed when
trying to add a custom filter on a datetime field.

It would be nice to add a test for this, but the code of the
SearchView cannot be easily tested yet.