[FIX] point_of_sale: recompute bank statement balance before validating session
Before this commit, the return flow in the pos would not work if closing the pos session intervened before the return payment registration.
The balance of the the statements was not updated before going to the validate step.

After this commit, we force the recomputing of the relevant fields (i.e. balance_end_real), to reflect reality and permitting validatioon afterwards

Backport of e27a3b185e449139660f43fea34de9a79dc66f2c (on #17438)
Closes #17599
OPW 743763