[FIX] calendar: _get_recurrent_date_by_event
Let's create in google calendar a reccurent event with all day
set and a end date.

When synchronising google calendar with Odoo(by clicking on button
"Sync with Google"). It raised on error because the start date had
a time zone but not the end date(the UNTIL in the rrule).

Fine tuning of 20ed57f874176161a043dee834b25c7f0b3dc175

opw:746549