[FIX] web: failed forward-port of https://github.com/odoo/odoo/commit/c4c9347b9986f60f1b8a534b805e248b25e3eef8
The mentioned commit exposed a `Dialog` instance thanks to a return and
an event triggering. The forward-port failed in saas-14 as the code
landed in the wrong function.

As the event trigger seems to be useless code without any context and
that this code has been hugely refactored in recent versions, this
commit simply removes it.
Feel free to expose variables again in version 11.0.

opw-748027