[FIX] mrp: set default quantity
The wizard to change the produced quantity of a manufacturing order was not
using the current produced quantity as the default value.
This was because the field mo_id was not present in the view (and not in
`fields` in default_get).

Adding the field as invisible to compute the default value for mo_id.

Correct typo in dictory lookup (using `[` instead of `(`)

Closes #17937
opw-749725