[FIX] *: add color option on m2m_tags widgets in kanban
There is a bit of magic in the kanban with many2many: the widget many2many_tags
is automatically set on many2many fields.

Since the previous commit, the `color_field` needs to be explicitly stated
in the options for the tags to be colored.

In kanban views where tags were previously colored, we thus have specified the
widget and the color option.