[IMP] sale_timesheet: projects of sale order
The stat button now shows all the project involved
in the sale order (created, or linked).

This commits introduce the field `project_ids`,
containing all the project involved in the sales
order (created, or linked via products).

The stat button now shows all involded projects (if
many) or the timesheet overview of the single project.