[FIX] web: properly display id in some cases
When a one2many is set through an onchange, the web client did not set
the id field properly.  Because of that, it could not be displayed in a
list. This is due to the fact that the id is given in the command like this:

[1, 53, {...}]

Here, the command is 1 (update), then 53 (id), then some values, which
do not have the id. So, in this specific case, we assumed that all the
new values were in the object, even though we properly set the res_id
property.

We fix this issue at the root: in the _makeDataPoint method, we make
sure that no datapoint can be created with a res_id and without the
corresponding id key.

Note that this issue was not found in a standard installation of Odoo
(but it may be present somewhere), but in some custom code.  It is rare
to actually use the id field in a view, but in this case, we had a
many2many in kanban view, and the id was used to format the url of an
image.