[FIX] Close POS session with amount 0 on statement lines
When trying to close a POS session (which confirms the bank statements
which reconciles the statement lines) if one of the statement lines
has an amount of 0 no payment object is
created (addons/account/models/account_bank_statement.py:905) and thus
``payment`` remains set to ``False``. In the past this was not an
issue but since b17e4a0f6b0af508e3e3f06b84aa43687bb8cae6 at the end of
the method (line addons/account/models/account_bank_statement.py:1003)
we unconditionally write to ``payment`` which blows up when it's
``False`` rather than a recordset.

This can happen if e.g. a POS order is composed only of free
products (don't ask...)

Fix by using an empty recordset instead of a boolean for a
non-payment, making payment.write a no-op.

OPW-752922