[FIX] website_portal_sale: remove incorrect field dependency
The field 'payment_acquirer_id' is added on the sale.order model
in the website_sale module, which may not be installed when this
code runs since it is higher up in the dependency tree.

Either way, it's quite convoluted to go fetch the acquirer on the
quote since it's already available on the transaction itself anyway.

opw-755998