[FIX] purchase : procurement check fail if one move in cancel
Use case to reproduce the bug:
One procurement is running for a PO. Cancel one move in the picking and
create it again. Pass the picking in done.
--> The procurement stay in running state

It happens because the check method of procurement look if all moves linked
are in done state but in this case, there is one move in cancel and one
move in done.

This commit check instead if all move are in cancel or done and at
least one is in done state.

opw-753177