[FIX] sale_stock: Adapt test due to c854de4
The tested use case is the following:
- Create a SO with 5 units of a product (invoiced on delivered quantity)
- Ship 5 units
- Create an invoice with 5 units, validate
- Return 2 units from delivery

The current state:
- Ordered : 5
- Delivered : 3
- Invoiced : 5

In that case you have invoiced more than the delivered quantity and the sales order should be marked as fully invoiced following the commit c854de4

The salesman will transfer the credit note to the accountant and everybody's happy.

As it is the intended behavior, this commit modifies the test.