[FIX] stock : return serial number
When an operation has the check for create new lots/serial numbers
activated and it exists a return that use this operation. The return
with serial number is not possible since it will display the view for
creating new serial number while we are returning some of them.

Trying to add the serial number name will fail due to the constraint
that ensure maximum one specific serial number by product.

This commit change the context before calling the view in order to
always display the many2one lot_id if the move is a return.