[IMP] sale: add name_get and name_search on SO
The goal is to have a name_get for SO returning
a value like 'SO0002 - Michel F.'

To avoid performance problem, the chosen solution
is to add a context key 'sale_show_partner_name' to
provide the partner name in name_get(), otherwise
it will always add it and kill the performance.
This solution is preferred over adding a stored
related field for partner name (trigger recompute
when partner is renamed for all SO is bad).