[FIX] mail: smooth scrolling in discuss for iOS app
In the iOS app, the content immediately ceases to
scroll when you remove your finger from the touchscreen.

Now, he content continues to scroll for a while after
finishing the scroll gesture and removing your finger
from the touchscreen.
The speed and duration of the continued scrolling is
proportional to how vigorous the scroll gesture was.

This Css property is the only way we find to have
the same behavior in the android app and in the iOS app.