[FIX] stock: move name length
- Create a stockable product with a long description (> 3000
  characters)
- Validate the SO

An error is raised because the stock move name is too long.

The stock move name is taken from the procurement name, which is itself
taken from the SO line description. However, the stock move name is an
indexed field and therefore its length is limited (to 2712 characters by
default).

We truncate the move name to 2000 characters. Actually, there is no real
point to so many characters, nor its indexing.

opw-767040