[FIX] stock: show_details view error on edit button
Show details view should directly relfect the stock move line
related to a move. However the user was not able to update
locations and it still exists a button that allow to edit
the move line. Since the move line is the final object everything
should be directly editable and the button should not exist

This commit update the view in order to be able to edit information
directly on it.