[FIX] web: fix kanban card color
Since the new kanban design, the cards' colors are shown as a
border-left color. Since all the cards have a small gray border,
a combination of inset box-shadow and border-left was used to display
the colored left border. There were two bugs with this system:

- The border-left was overflowing the box-shadow of 1px, making the
  border look weird
- The box shadow was always below the content (graphs, progressbar, etc)

The first problem is resolved by leaving the border-left property
untouched (the natural border is not really visible anyway), the second
problem is resolved by using the ::after element to create the border
instead of using a box-shadow.