[FIX] purchase: Import res_config_settings.py after res_company.py
Purpose
=======

Steps :
- Blank DB (without demo)
- Install inventory, Sales
- Create some products + update qty + activate serial number from setting and create some products (I tried this step and faced traceback)
- Now try to install Purchase app
It gives traceback..

Specification
=============

As the field po_lock is defined on the res_company and a related field is defined on the res_config_settings.py file, modify the import order.

change the position of res.company file, so that fields of po_lock will define first and then it will not generate field's error in res.config.settings file.