[FIX] web: skip buttons when navigating in editable list
Clicking on a cell that contains a button (but not ON the button) leads to
a traceback (see "Scrap" button in Stock Operation for example).

This is due to the fact that a button is not registered as a widget so it can't
be activated.

This commit ensures that the next line is selected if the last column is a
buttton.