[REF] stock: move compute_state logic to stock move file
Before when comuting state picking there was a computation
among all moves in order to guess the picking state.

However this code could be reuse in other scenario for
example in merge moves.

In order to avoid code duplication this commit move the
part responsible for the state computation among moves
in a specific method inside stock move file.