[IMP] mass_mailing: consider replying to an email imply opening it
Currently when people answers an email its statistics are set to replied
updating its replied field. However if the blank gif used to track the
opening fails for whatever reason we could have mail considered as replies
but not opened.

We consider that it is not common to reply to an email people did not
open. Let us therefore set a replied email as also opened.