[FIX] web, *: review kanban record sizes
* account, maintenance, mrp, payment, point_of_sale, sales_team

Previous system was:
- `flex: 1 1 300px;` on kanban records
- If a specific record needs a different size, add custom style to
  either change the `flex` rule or set a `min-width` for >=SM screens

Now:
- `flex: 1 1 auto;` and `width: 300px;` on kanban records
- If a specific record needs a different size, add custom style to
  change the `width` rule.

This allows some standardization of the way to customize the suited
width and also allow lesser LESS code (as the previous version required
either the use of the flex mixin or the use of a media query).

Note:
- Also remove useless app record rule
- Also fix MRP Work Centers record width

Note2:
This system should be improved for version 12.0.