[FIX] stock: scrap creates a back order.
When the user scrap a product in a picking.
The moves to proccess are moved to a back order and
the current picking only contains the scrap move.

It happens because scrap use action done on the scrap
move and action done process the given move and put the
other in a back order.

This commit add a condition on back order creation in
order to not do it if it comes from a scap.