[IMP] mrp: only show reserved when no show operations on picking type
Because the stock.move.line view (when Show Operations is checked on the
picking type) does not take into account the show reserved setting on
the picking type, it is better not to show the checkbox when show
operations is checked. We add an onchange to always check the "show
reserved" settings when "show operaions" is checked.