[FIX] web: max one wiggle per readonly click
The wiggling effect of the edit button when clicking on the form
(hinting that to change the form you should edit it) could  happen
several times.

With 85d1e76 there was also a visual deffect on the Edit button when the
wiggle happened several times, because bounce are stacked but the CSS
class was removed after the first bounce was removed.