[FIX] website_sale(*): better management of warning when stock no more available + ...
From now update_json route with display=1 return the cart_qty and not False.
It is usefull withtout extra cost for other modules like wishlist or ... that
previously render all the template just to know the new qty.

Now you cannot arrive on payment page withtout product in your cart and get a
blank page, you are now redirected to the /shop/cart page

When the order are altered due to a change of stock availibilities, we now log
it into the order_line itself or the saloe_order if a line has been deleted.
That permit to display the warning later when the user come back to one of the
page that display the cart resume.  (/shop/cart or /shop/payment at this moment).

Remove useless empty file

Make xpath correctly using wrap feature and don't re-copy all contents

Replace old base.get_context with web_editor.context