[FIX] stock: put in pack with UOM
When put a quantity in a pack the reserved quantity is
split between the original move line and the move line
for the pack.

However in order to compute the quantity left to do. The
system compare the quantity reserved in term of quant UOM
and the quantity done with the move line UOM.

This commit compare both fields in move_line's UOM.