[FIX] web: calendar IE9+ incompatibility
The "Event" API is not the same between IE9 till IE11 and all the other
browser (eg: IE Edge, google chrome, firefox, ...).

For example here is a seemingly accepted polyfill to have it working the
'same' on most browsers using CustomEvent:

https://developer.mozilla.org/en-US/docs/Web/API/CustomEvent/CustomEvent#Polyfill

This commit just trigger the event with jQuery which attains the same
result but work everywhere.

opw-780056
closes #20759