[FIX] project, sale_timesheet: analytic account and project creation
When we create an analytic account, a project is automatically
generate, and it should not. This is a bug coming from
the removal of 'use_tasks' field. See eb4ad335137de7831902e760e3fc375eeb81f326.

We can now remove all the 'project_create' and
'project_creation_in_progress' context key.