[FIX] mrp: in move validation, take into account the owner
Before this commit, the special mrp validate_move method did a call to
quants_move with an incomplete signature. IN the use case, that we want
to make "own" produced quantities by a third party (maquila), we can do so in
downstream modules by restrict the partner on the move so that any moved (or
created) quants will belong to this partner (owner).  This is certainly
important for the case where finished products are not your own
property.  Courtesy of blaggacao.

Closes #21072